Added New Tables for Stock and Mutual Fund transactions Added new tables for stock and mutual fund transactions. Updated the integration test to automatically copy and test the sql file.
