+++ user-data/src/test/java/com/mydata/userdata/integration/InvestmentTest.java
@@ -13,65 +13,21 @@ import com.mydata.userdata.dto.MutualFundDto;
 import com.mydata.userdata.dto.StockDto;
 import com.mydata.userdata.utils.GenerateTestPojo;
 import com.mydata.utilities.test.conroller.ControllerTest;
-import io.r2dbc.spi.ConnectionFactoryOptions;
 import java.util.Objects;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.mockito.junit.jupiter.MockitoSettings;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
-import org.springframework.test.web.reactive.server.WebTestClient;
-import org.testcontainers.containers.PostgreSQLContainer;
-import org.testcontainers.containers.PostgreSQLR2DBCDatabaseContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
 
-@MockitoSettings
 @SpringBootTest(
     webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT,
     properties = "server.port=0")
-@Testcontainers
 @SuppressWarnings("PMD.BeanMembersShouldSerialize")
-class InvestmentIntegrationTest implements ControllerTest {
+class InvestmentTest extends IntegrationTestBase implements ControllerTest {
 
   public static final int MIN_RECORDS_RESULT = 5;
-  private static final String POSTGRESQL_IMAGE = "postgres:13.5";
-  private static final String DB_NAME = "testDB";
-  private static final String USER_NAME = "testUser";
-  private static final String PASSWORD = "testPassword";
-
-  @Container
-  private static final PostgreSQLContainer<?> postgresql =
-      new PostgreSQLContainer<>(POSTGRESQL_IMAGE)
-          .withDatabaseName(DB_NAME)
-          .withUsername(USER_NAME)
-          .withPassword(PASSWORD)
-          .withInitScript("sql/db-init.sql");
 
   private final BeanTransformer dtoToDtoSkipId =
       new BeanUtils().getTransformer().skipTransformationForField(ACCOUNT_ID);
-
-  @Autowired private WebTestClient webTestClient;
-
-  @DynamicPropertySource
-  @SuppressWarnings("PMD.DataflowAnomalyAnalysis")
-  public static void properties(final DynamicPropertyRegistry registry) {
-    var options = PostgreSQLR2DBCDatabaseContainer.getOptions(postgresql);
-    registry.add("db.driver", () -> options.getRequiredValue(ConnectionFactoryOptions.DRIVER));
-    registry.add("db.host", () -> options.getRequiredValue(ConnectionFactoryOptions.HOST));
-    registry.add("db.port", () -> options.getRequiredValue(ConnectionFactoryOptions.PORT));
-    registry.add("db.user", () -> options.getRequiredValue(ConnectionFactoryOptions.USER));
-    registry.add("db.password", () -> options.getRequiredValue(ConnectionFactoryOptions.PASSWORD));
-    registry.add("db.database", () -> options.getRequiredValue(ConnectionFactoryOptions.DATABASE));
-    registry.add("db.schema", () -> "public");
-  }
-
-  @Override
-  public WebTestClient getWebTestClient() {
-    return webTestClient;
-  }
 
   @Override
   public String getBaseUrl() {
