+++ user-data/src/test/resources/sql/db-init.sql
@@ -1,69 +1,3 @@
--- Create All Tables - START
-create TABLE IF NOT EXISTS deposit_account(
-    id serial PRIMARY KEY,
-    bank_name text NOT NULL,
-    branch text,
-    account_number text NOT NULL UNIQUE,
-    balance numeric(12,2) NOT NULL,
-    created_at timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    modified_at timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    active boolean NOT NULL DEFAULT TRUE
-);
-
-create TABLE IF NOT EXISTS loan_account(
-    id serial PRIMARY KEY,
-    bank_name text NOT NULL,
-    branch text,
-    account_number text NOT NULL UNIQUE,
-    balance numeric(12,2) NOT NULL,
-    created_at timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    modified_at timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    active boolean NOT NULL DEFAULT TRUE
-);
-
-create TABLE IF NOT EXISTS miscellaneous(
-    id serial PRIMARY KEY,
-    investment_name text NOT NULL UNIQUE,
-    balance numeric(12,2) not null,
-    created_at timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    modified_at timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    active boolean NOT NULL DEFAULT TRUE
-);
-
-create TABLE IF NOT EXISTS mutual_fund(
-    id serial PRIMARY KEY,
-    mf_code text NOT NULL UNIQUE,
-    mf_name text NOT NULL UNIQUE,
-    amc text NOT NULL,
-    type text NOT NULL,
-    created_at timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    modified_at timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    active boolean NOT NULL DEFAULT TRUE
-);
-
-create TABLE IF NOT EXISTS saving_account(
-    id serial PRIMARY KEY,
-    bank_name text NOT NULL,
-    branch text,
-    account_number text NOT NULL UNIQUE,
-    balance numeric(12,2) NOT NULL,
-    created_at timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    modified_at timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    active boolean NOT NULL DEFAULT TRUE
-);
-
-create TABLE IF NOT EXISTS stock(
-    id serial PRIMARY KEY,
-    stock_code text NOT NULL UNIQUE,
-    stock_name text NOT NULL UNIQUE,
-    stock_exchange text NOT NULL,
-    broker text NOT NULL,
-    created_at timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    modified_at timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    active boolean NOT NULL DEFAULT TRUE
-);
--- Create All Tables - END
-
 -- Insert Deposit Account records
 insert into deposit_account (bank_name, branch, account_number, balance) values ('Bank Name One', 'Branch One', '1', 1.1);
 insert into deposit_account (bank_name, branch, account_number, balance) values ('Bank Name Two', 'Branch Two', '22', 22.22);
